Feat: 홈 화면 - 플로팅 알람 설정
