+++ src/pages/home/index.tsx
@@ -8,7 +8,6 @@ import SpecialServices from "./components/SpecialServices";
 import CategoryList from "./components/CategoryList";
 import SubServices from "./components/SubServices";
 import Footer from "./components/Footer";
-import FloatingAlarm from "./components/FloatingAlarm";
 import { Suspense } from "react";
 import { useEffect, useState } from "react";
 import useFCMToken from "./hooks/useFCMToken";
@@ -16,6 +15,9 @@ import { requestPermission } from "../../firebase-messaging-sw";
 import Cookie from "js-cookie";
 import { getStoredToken } from "@utils/indexDB";
 import useNotifications from "@pages/notice/queries";
+import Lottie from "lottie-react";
+import approveLottie from "@assets/lotties/approve.json";
+import { Link } from "react-router-dom";
 
 const Home = () => {
   const isLoggedIn = Cookie.get("isLoggedIn") === "yes";
@@ -52,6 +54,15 @@ const Home = () => {
     }
   }, [isSuccess]);
 
+  const tradeRequest = data?.data.notifications.find((item) => item.type === "TRADE_REQUEST");
+  const clickedNotification = localStorage.getItem("notificationId");
+  const checkNotification =
+    tradeRequest && tradeRequest.notificationId.toString() !== clickedNotification;
+
+  const handleNotificationClick = () => {
+    localStorage.setItem("notificationId", tradeRequest?.notificationId.toString() || "");
+  };
+
   return (
     <>
       <S.Container>
@@ -82,8 +93,21 @@ const Home = () => {
         </Suspense>
         <Footer />
       </S.Container>
-      {data?.data.notifications.map(
-        (item) => item.type === "TRADE_REQUEST" && <FloatingAlarm key={item.notificationId} />
+      {checkNotification && (
+        <Link to="/mypage/management" onClick={handleNotificationClick}>
+          <S.AlarmContainer>
+            <S.LottieContainer>
+              <Lottie animationData={approveLottie} />
+            </S.LottieContainer>
+            <S.TextContainer>
+              <S.LinkContainer>
+                <S.LinkText>판매승인하러 가기</S.LinkText>
+                <S.ArrowIcon />
+              </S.LinkContainer>
+              <S.Text>등록하신 매물이 승인 대기중입니다.</S.Text>
+            </S.TextContainer>
+          </S.AlarmContainer>
+        </Link>
       )}
     </>
   );
